Add a dirty bug fix
The problem occurs when we allocate from the unaligned chunk. In some
cases the line "left_size = largest_contained_power_of_2(current->size)"
causes some problems, since it chooses a two big size for left chunk
and leaves too less space for the right chunk. And, since afterwards the
right chunk becomes the current chunk we return less memory that we
should. For now i fixed the problem by ignoring the unaligned initial
chunk, but of course, this is a temporary solution.